sim: parsim ported, see details below
Most changes come from the need to follow the signature change of
cGate::deliver(), especially the addition of SendOptions as parameter.
Due to that, whenever a cMessage is sent from one partition to another,
it needs to be acompanied by the SendOptions. Thus, SendOptions needs
to be packed/unpacked, etc.