sim: cModule: change module destruction sequence
So that when deleting a compound module, the compound module's local
listeners are notified about the deletion of the submodules.

The downside is that more listeners now need to be explicitly unsubscribed
(listeners MUST NOT be subscribed anywhere when they are destructed,
cIListener's destructor contains an assertion for that). Due to this change,
signals_modelchange_1.test currently breaks (error during network teardown).