qtenv: Show transmissions on ideal channels as a full-length line
This is useful when a transmission is sent on a connection path
that has both a transmission channel and some ideal channels.
No holding (zero-time) animation (of it growing/shrinking) yet,
because we wouldn't have an exact opportunity to request the hold
at the right time (e.g. when the front of the transmission passes
from a transmission channel into an ideal channel).