sim: cComponent: added preDelete()
It can help simplify network or module deletion in a complex simulation
that involves model change listeners. Assume that a model change
listener in a module contains code that involves accessing other modules
(e.g. calling methods in them). Normally, when those other modules
don't exist, the code would throw an exception to indicate the error.

However, when the same listener is invoked due to module deletion,
those collaborator modules may no longer exist (just having been deleted).
This can result in a crash; but even when the nonexistence (disappearance)
of those modules is properly detected, the code is not supposed to throw
an exception -- which is different from its normal behavior. Writing
code that handles such situations can be tricky.

The preDelete() method can help here: if it is overridden to
unsubscribe from the model change listener, the above complications
can be avoided altogether.