sim: cModule: deleteModule() refactoring
It is not supposed to do stuff (like fire pre-model-change notifications)
from a halfway-deleted module. Now we do every risky thing (such as
deleting submodules and disconnecting gates) from doDeleteModule(), and
only delete the module object when it is already barebones (no
submodules, gates, listeners, etc).

The deletion sequence is now pretty much the reverse of the setup
sequence.