build: do NOT use --rpath on Windows at all as it is meaningless and LLD does not support it at all.
Note that --rpath is autodectected doring configuration, but the standard
mingw windows linker accepts it however, it is a NoOp. We rather not use
it at all on Windows because LLD on Windows explicitly errors on --rpath.

Previously, rpath was disabled only for the MS-ABI compatible linker, but it's better
not to use it at all.