build: use Group Target in generated makefiles when creating shared library
On Windows, we have to generate both a .dll and an export lib (.dll.a) when creating a shared lib.
These two files are created by the linker at the same time, so we have to specify
that they are not separate dependencies, but they are generated by the same recipe,
otherwise the linker would be started two times when using parallel build.

Sadly the &: syntax is supported only on gnu make 4.3+, but luckily we bundle
that version on Windows (where we need it). On macOS and Linux (where only gno make 4.2.1 is available)
it silently falls back to the original : behavior (i.e. simple dependency).
But on these platform we have only a single target anyway (no importlib is generated)
so it does not matter.

NOTE: this might still cause the above issue if cross compilation to Windows is used
on a Linux host, while the installed make is still 4.2 or lower. It's not a critical bug,
it just makes the build process longer by unnecessarily running the linker twice.