sim: remove enable-auto-import flag
Using auto-import on Windows emulates the behavior of Unix systems where all
functions are implicitly exported. This flag however can result in
exporting too many functions and in certain projects (e.g. INET) can exceed
the 65535 hard limit causing serious crashes.

From now on, all classes meant to be exported as an API *MUST* be marked
by PROJECTNAME_API macro that is expanded to OPP_DLLEXPORT or OPP_DLLIMPORT
depending whether the project is currently generated or used by
a 3rd party project.

NOTE: you can emulate the explicit export requirement on Linux
by adding -fvisibility=hidden to the CFLAGS. This allows testing on Linux whether
all APIs are correctly marked to be exported. To support this, the
OPP_DLLEXPORT macro is also used on non-Windows systems where it is expanded
to a visibility(public) attribute.