ide/launch: Fix printing to the console of processes right after their launch
It has to be done asynchronously if it cannot be done right away
because there might not be a console allocated for them yet.
And I found no way to "force-allocate" them synchronously.
Also there is no need to allocate a new output stream each time,
we can simply "hijack" the stdout and stderr streams ... I hope.