qtenv: Change the factory default of animating methodcalls to "off", rename the config key
The latter is needed for this change to take effect even for existing installations.
Reason for changing the default:
In INET all Enter_Method_Silent calls were replaced with Enter_Method,
and this resulted in lots of unnecessary animations.
Method call animations are now seen as debugging tool, or aid for understanding
C++ code; showing a lot of low-level technical details, so it should be off by default.
There are better visualizations built into INET anyways.