ide/scave: fix: Triangle/Tri symbols on the legend didn't reflect symbols' orientation on the plot
Solution: Marker symbols with different orientations are made
subclasses. This ensures that the class name alone, which is
used as a key to the image cache, describes the shape.

This makes sure they will have a different image cache key when
they are drawn into the legend tooltip, so the first orientation
doesn't get "stuck" in there.