sim: cObjectParImpl: forEachChild() should not be a state-changing operation
An objectValue() call, if the parameter contains an expression,
would re-evaluate the expression and delete the previously returned
object while storing the new object into 'obj'. This crashed Qtenv's
Object Inspector, because it tried to use the old pointer (namely,
tried to get a cClassDescriptor for it), not suspecting that
collecting the object's children via forEachChild() would delete the
object.

The solution is basically that forEach() doesn't recurse into the
value produced by volatile parameters, which kinda makes sense
(they are volatile).

Same for cXMLParImpl.