ide/inifile: line parsing adapted for multi-line entries
Main issue is how to handle value+comment separation with
multi-line values. There doesn't seem to be meaningful
to edit the "pure" value and the comments separately,
because the comments may contain commented-out lines,
comments specific to some part of the value etc, which
may become misaligned when editing separately.

The only feasible solution for multi-line values is to leave
the comments as part of the value, and let the user them
as one string. (For consistency, the 'comment' part of the
entry is then set to "".)

For validation etc, the net value is needed (ie stripped of comments).

Therefore, we have now getRawValue()/setRawValue() in addition
to getValue()/setValue().

For single-line values, nothing changed. getRawValue() returns
the same thing as getValue().