sim: NED: implemented @omittedTypename property + test
@omittedTypename allows one to specify a NED type to use when
typename="" is specified for a parametric submodule or channel,
or to remove (skip creating) that submodule, submodule vector element,
or channel altogether.

@omittedTypename can be specified on a module interface or channel
interface, or on a submodule or connection. It should contain a
single (optional) value. The value names the type to use. If it is
absent, the submodule or channel will not be created. (The connection
will be created without a channel object.)