sim: Add proxy class descriptor for WATCH_[OBJ|PTR] to avoid crashes
The original solution involving:
  virtual cClassDescriptor *cWatch_cObject::getDescriptor() const override {return r.getDescriptor();}
didn't work because the cWatch subclasses lied that they themselves can be described
by the descriptors of the watched objects, which is not true.

And since there is often a cast through a void* in the descriptor APIs,
this mismatch went unnoticed, and operated essentially as a reinterpret_cast,
so mostly garbage was shown on the GUI as field values, until eventually a crash occurred.

See test/anim/watch for an example.