sim: NED: allow the body of a parametric submodule to assign apparently nonexistent parameters
Example:

        app: <default("FooApp")> like IApp {
            parameters:
                address = parent.address;
                sleepTime = 1s;
        }

If FooApp has a sleepTime parameter but IApp does not,
then running with the defaults is fine, but changing to
another IApp type that does NOT have a sleepTime parameter
will stop the simulation with an error.

Solution: inside a parameteric submodule type, we simply ignore
parameter assignments that assign nonexistent parameters.