sim: assume cSimpleModule to reduce calls to dynamic_cast
On windows, approx 3% of execution time is spent in `cMessage::isStale()`.
Largely this is because of the dynamic_cast. Since the isTerminated
check is actually a trivial bit comparison it can be done before
dynamic_cast which is relatively slow. Then the dynamic_cast is done
only if it evaluates as true.