ide/scave: Implement equals() and hashCode() in DataTable.Column properly
Because now there are two columns with the text "Value", but they are
not actually the same. As a result, the scalar values were missing.
The default* fields are left out from both methods, because they
can change based on user preferences, so they shouldn't contribute
to the identity of the column.