ide/scave: Add support for x:id.enablerButton attribute in XSWT files
If control is something (e.g. a Text) that should be enabled conditionally
(e.g. when a radiobutton or checkbox is selected), specify the ID
of the controlling button in the x:id.enablerButton attribute of the control.

Also fix: for x:id.isEnabler=true, if the button is a radiobutton, it should
not disable its sibling radiobuttons.