ide/scave: Remove busy UI event loops from BrowseDataPage.doRefreshPage()
This fixes a livelock that occurs (mostly on Windows) on startup,
when the active editor tab is a chart.
And they are most likely not needed anymore, because the process whose
progress they were meant to display has been optimized a lot lately,
so it should finish quickly most of the time.