tests: fixed a negative coroutine test failure on Windows
Using the Windows Fiber API, coroutine stacks do not have to be pre-allocated,
so this test was meaningless. Fiber creation is limited only by the total amount
of physical memory, so the test never resulted in an error.

When using the FIBER API, the test is now skipped.

The test got also finetuned, so it requires less memory. Just skipping the test
on Windows required the whole network to be instantiated and that requires a lot
of time and memory.