scave: IMPORTANT: Make sure we only read vector files that correspond to the loaded state
We must NOT load vectors from vec files that do not correspond to the
state loaded in ResultFileManager. Those vec files might have been
re-generated with entirely different content.

We use the fingerprint for that. (Might use runId as well, but we want
to disallow appended vec files too, because their stats do not
correspond to that displayed on the Browse Data page. So runId is
not good enough.)

This change affects the Vector Browser view, and the Python "results"
API as well (indirectly, via ResultsProvider and ResultsPickler).

Handling of such errors (exceptions) is not nice, it's rectified in the
next commit.