ide/NED editor: fix #634
Eliminate NED file launcher class (which opened either a NED editor or a
plain text editor, depending on whether the NED file is part of a
NED folder or not), and instead make the normal NED editor capable of
handling files outside NED folders.

This makes clicking in the Search Results view work as advertised.