sim,envir: START EXTENSIVE REFACTORING OF ENVIR
Goal: Make it possible to use OMNeT++ from Python as a library,
with the help of the cppyy package [1,2].

[1] https://github.com/wlav/cppyy/
[2] https://cppyy.readthedocs.io/

This commit adds an initial working Python script that loads
OMNeT++ libraries and runs the Aloha example simulation.

An additional goal is to allow the use of multiple theads/cores
for running simulations. Each simulation would be STRICTLY tied
to one thread, but the use of multiple threads would be allowed.
This is largely achieved by having per-thread copies of global
variables, via the use of the "thread_local" C++ modifier.

Refactoring involves:

- Breaking up EnvirBase, Cmdenv, Qtenv into library classes which
  are usable as building blocks from Python;

- Make multiple instances of cSimulation possible, and even, the norm.

- cEnvir/cRunnableEnvir/EnvirBase now serve a dual role: "the thing
  that provides services for cSimulation", and "the app that sets up
  and runs simulations". These two need to be separated, and the latter
  functionality factored out into distict classes. cEnvir and subclasses
  should purely be a class that plugs into cSimulation and provides
  services (such as recording simulation results) for it.

- Getting rid of singletons and global state as much as possible.
  (Note that it is not possible to fully remove globals, as it would
  imply changing the simulation kernel API in a way that would
  severely inconvenience model authors.)