common: expression.y: eliminate sanitizer warning
Warning came while checking if a double value can be losslessly represented in int64_t.
Code was legit as it was, but now the sanitizer is happy too.

Sanitizer report:

expression.y:402:42: runtime error: 1e+200 is outside the range of representable values of type 'long'
SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior expression.y:402:42