prod: added 'fp-contract=off' in release mode to prevent FMA (fused multiply-add) optimization.
This is necessary because:
- In debug mode, -O0 is used which turns off FMA, so we may get slightly different floating
  point results in debug and release mode which could result in different
  simulation trajectories and fingerprints.
- Even in release mode, refactoring the code may result in different results
  because how and when the compiler applies FMA optimization is unpredictable.
  This is rather dangerous for the simulation models.

We prefer predictability over speed and accuracy.
NOTE: it is OK to turn this on for parts of the code using pragmas.

The default of 'fp-contract=fast' has changed to 'fp-contract=on' in clang 14,
so we have to disable this manually.
See: https://clang.llvm.org/docs/UsersManual.html#cmdoption-ffp-contract