nedxml: ned2 lexer/parser: modernization, made reentrant
modulo the "parenDepth", "pos" and "ps" global variables

NOTE: In ned2.lex, I tried to replace the count() solution with
the YY_USER_ACTION based solution on stackoverflow.com, but
couldn't not get it to work due to artifacts in the lex file:
- use of yyless()
- some rules require extend() and not count()

In addition, the resulting code would be even more messy and
hard to understand than the current one.

It is preposterous that Flex STILL doesn't support such a basic thing
as location tracking (bison's yylloc) out of the box, and one has
to resort to ugly hacks to achieve it.