ide/neddoc: added support for "@debug links" page directive
When a page (comment) contains "@debug links", NED/msg types will be
hyperlinked in a way that also shows the original code.

E.g. "~inet::Foo" in the page will not simply be rendered as a link with
the text "Foo", but like this: "[~inet::Foo -> Foo]" (where the last
"Foo" is the link).

This is useful for the test in test/misc/neddoc.

# Conflicts:
#	ui/org.omnetpp.neddoc/src/org/omnetpp/neddoc/DocumentationGenerator.java