qtenv: fix+refactor: Make messageDuplicated() return bool, move it down into MessageAnimation
This then fixes the crashes due to double-inserted pointers in the cache!

The same (original, used by the model) cMessage object can be sent
multiple times (like in the routing sample) during its lifetime, and
its multiple "sendings" can even be animated concurrently (for example
in the cut-through config of the routing sample).

Until now, the MessageAnimations that animated earlier sends of the
same message already had their private duplicate pointers, and didn't
care about the new private duplicate, and it didn't matter.

But this way, we inserted multiple entries into the animationsForMessages
map, even for animations that didn't care about the new duplicate.

This resulted in more entries than necessary, even invalid ones, and then
eventually dangling pointers.

This change only inserts new entries to the messages-to-animations map
if the notified animations actually took the new duplicate "as their own".

And since we already know which animations might be potentially interested
in the duplication of any message, by looking them up in the map, which
are MessageAnimation pointers, not plain Animation pointers, we don't
need the messageDuplicated() method to be declared in Animation animore.