test: fix incorrect expected output in inifile_eventlog_1.test
Events at t=2s and t=6s should be missing from the recorded eventlog file,
because the 1s..3s and 5s..7s intervals are excluded by the

  eventlog-recording-intervals = ..1, 3..5, 7..

line in the ini file.

NOTE: this causes the test to fail, exposing a bug which will be
fixed in the next commit.