sim,envir: loggingEnabled flag moved from cEnvir to cLog
Because it logically belongs there, with the other cLog static members
like logLevel, noncomponentLogPredicate and componentLogPredicate.

Slight con: it is thread-local but not per-Envir, so, when switching
between simulations in a single thread, now cLog::loggingEnabled needs
to be swapped too. (However, the same has always been true for cLog::logLevel,
so things didn't really become worse in that regard.)