build: rename WITH_PYTHON to WITH_PYTHONSIM
Reason: WITH_PYTHON specifically controls the availability of
Python-specific simulation features, such as the pyeval() and pycode()
functions, and the NED @pythonClass module attribute.

The name WITH_PYTHON was confusing: even if WITH_PYTHON=no was set,
we still required Python, as command-line tools are written in Python,
charts in the IDE Analysis Tool are Python-based, etc.