ide/scave: improvement: Reload user chart templates periodically in releases as well
This functionality was originally intended to be useful for development,
but now that charts can be exported as templates, it's better to not
have to reopen the editor to be able to use the newly exported ones.

In debug mode, all templates are still reloaded; while in release mode,
only those in the user's projects are.