qtenv: improvement: Perform holding animations in pausePoint()
This keeps it consistent with how it behaves when Stepping through
events with pausePoints one by one (using F4). That means, animations
for messages sent in an event before a pausePoint() are neither lost,
nor bunched together with send animations produced after the
pausePoint() (in case "broadcast animation" is enabled).