ide/neddoc: workaround for interfaces not being resolved when NED file was open in the editor
If a NED file is open in the editor when the IDE starts, and was not
edited, and it contains a module type which implements a NED interfaces,
the interface type cannot be resolved. This result

Example:

In inet, see:
- module Ieee80211Mac extends MacProtocolBase like IIeee80211Mac {...}
- in Ieee80211Interface's submodules:
      mac: <default("Ieee80211Mac")> like IIeee80211Mac

1. open Ieee80211Mac in the IDE, then close+restart the IDE
2. generate NED documentation for INET
3. open the doc, and go to the Ieee80211Interface page. In the graphics,
   hover above the mac module. The tooltip only shows the docu of
   the moduleinterface IIeee80211Mac but not of Ieee80211Mac which is
   the default type for the mac submodule

Reason:

NedTypeInfo of Ieee80211Mac's getInterfaces() returns empty set, could
not resolve IIeee80211Mac.

Workaround:

Create+use immutable INedTypeResolver which does a rehash() when it is created.