envir: fix "network not found" errors when INET was in a symlinked directory
NedResourceCache (or rather, the new canonicalize()) now resolves symlinks
when using dirs as map keys. This is necessary because it is impossible
to convert relative paths to absolute by preserving the symlinks in the
ancestor dirs, because getcwd() always returns a path with symlinks resolved.

Also appears to fix #747 "Module class not found due to soft links in the NED folder's path"