ide/scave: refactor: Chart: Replace `editedChartScripts` parameters with `Chart.editedScript`
This fixes a bug when copy-pasting a folder with a chart inside that
is currently open in an editor, and its script has been altered.
Previously, the newly created clone in the pasted folder had an outdated
script, now it always has the updated script text even if the analysis
file hasn't been saved yet.

And overally it looks a bit less terrible.