ide/scave: QuantityFormatterRegistry: save/restore contents as preferences, 1st attempt
Note that this was the wrong approach: the correct one is to set the
preference defaults prior to setting/getting the values on/from the
preference store.

Reason: The preference store only stores the values that are DIFFERENT FROM THE
DEFAULTS, which kinda makes sense. I.e. as the software evolves, a change in the default
value of a setting can take effect in the user's installation.