ide/scave: fix: all recorded parameters were listed with "type=double" in the Analysis Tool GUI.
Namely, in the data tree on the All tab of the Browse Data page.
We now don't show the parameter type, because that info is not
available (not currently recorded into the .sca files).

Clarify the meaning of ResultItem::getDataType(), and let ParameterItem's
method return NA to indicate that it is not meaningful for parameter items.