qtenv: fix: SendOnConnAnimation: Store source module ID on creation
Similar to how it is already in SendDirectAnimation.
This fixes a rare crash that happened when Qtenv was closed
with a message animation in progress on a degenerate connection.
In this case, gate->getOwnerModule() was a garbage value,
because the module destruction order (from the perspective of
message animations) is different than in the regular case.