sim: fix a mutex related crash on macOS
A mutex must be initialized before used. The static registerSignal call was
using the uninitialized mutex when the PRE/POST_MODEL_CHANGE signals were registered
in a static initializer block. When the mutex variable is in the static initializer
block in a shared library, it is not ensured that the mutex is initialized
before any other static variables. For this reason the static mutex variable
is placed at the beginning of the init block.