test/core: cDynamicExpression_unit_1.test: ignore result of "3km % 2m"
In OMNeT++ 4.0, '%' became an integer modulo operation (as opposed to fmod()-like).
Internally, numbers are represented with double, which are truncated
towards zero before the modulo operation is applied. In the example, 2m
becomes 0.002km, truncated to zero, so internally fmod(3,0) is done
which yields NaN on at least many systems. However, this is not a useful
expected behavior, hence the removal from the test.

From OMNeT++ 4.1 on, '%' only accepts dimensionless numbers, so the same
weirdness doesn't occur.