doc/usman: workaround: wrong viewbox for dot-generated SVG figures
See "Some figures in the Simulation Manual appear to be zoomed to their top-left corner"
https://github.com/omnetpp/omnetpp/issues/1048

Since the viewBox correction factor differs from figure to figure, this
fix is not perfect. factor=2.0 seems to be fine for most figures, and
only seriously off (too much) for one.

Both web and pdf renderings look ok.