ide/scave: Try to generate relative paths into the .anf file in the creation wizard
When opening a .sca or .vec file, make the input patterns as relative to
the .anf file under creation, unless they would have to start with "../".
This will leave out the project name in most cases, simplifying locating
the result files by opp_charttool a lot.
This was already done when adding inpuots to the analysis by dragging
and dropping .sca or .vec files into the inputs tree.