ide/scave: Don't rely on backend methods removed in Matplotlib 3.8
A couple methods from FigureCanvasBase related to event dispatching
were removed that we relied upon in the IDE backend.
Porting over to the new way of forwarding events.

Also removing wheelEvent and mouseEventCoords - these were unused
leftovers taken over from the Qt backend, which was used as
inspiration for the initial inspiration.