ide/inifile: ModuleHierarchyView: fix "Unhandled event loop" exceptions" previously worked around
workaround commit: 164f17ba71 ("ide/inifile: workaround for "Unhandled event loop" exceptions", 2023-10-03)

Error occurred when deleting a submodule while the Module Hierarchy view
was open. Cause was that during computing the icon for a submodule, it
wanted to find the containing compound module, and that submodule
didn't have one. It was a stray submodule. That should never have happened.

Root cause: After submodule is deleted, the NED resolver sent a change
notification, which properly triggered an AbstractModelView.rebuildContent()
call in the Module Hierary View. It properly called buildContent(),
which used GenericTreeUtils.treeEquals() to determine if viewer content
needs to be replaced with the new one.

However, it called GenericTreeUtils.treeEquals() with lazy=true, which ignores
children if they haven't been computed. Therefore, the old and the new NED
trees were regarded as equal, so the viewer's input was not replaced.
This led to the view trying to display the icon of the submodule
which was no longer in the NED file tree (only in the old copy).

Fix: Non-lazy comparison. Removed the lazy mode of GenericTreeUtils.treeEquals(), because
it is inherently dangerous to use. It was not used anywhere else in the code.

Removed previous workaround, too.