ide/scave: Change automatic line chart area computation logic
The main motivation was to make sure that data which is far from
the origin, but spans a relatively small range (say, 10000..10005),
is still plotted so that the points are not too close to the edge
of the axes area, making them hard to find/see.

This is done by calculating the padding after including the origin
in the bounds rectangle, so it will be sized relative to the "10005",
and not to the "5" in the example above.

This should be more intuitive.