build: do not use a temporary static library when building an exe
For long command lines (>8000char), a temporary static library was created
and that library was linked with an empty executable file. Unfortunately
this static library required the --whole-archive flag otherwise the
linker symbol garbage collector would throw out unreferenced symbols
(which is almost all the symbols in the model static library).

We avoid this issue by using the command option file directly on the linker,
instead of using it for creating a static library first and then using the
temporary library file on the linker command. This also speeds up the build
process as it avoids the creation of a big, temporary static library
as an intermediate step.