sim: cPar: fix issue #1132: parse() now restores the evaluation context to the owner component
Added resetEvalContext=true arg to parse().

It was chosen to add the restoring code into parse(), because it is
called at multiple places, and almost always the intention is that
the expression should be evaluated in the context of the component
that owns the parameter.

Also: added cPar::getOwnerComponent() convenience function.