qtenv: Fix logging issue when something is logged from scheduler context
The log messages were added to the previous event which had nothing to do with the scheduler callbacks.
This behavior was very confusing in INET simulations with ExternalProcess that log to their stdout/stderr.
See the showcases/emulation/webserver example.

Changed Qtenv log window to add a separate banner for log messages recorded from the
RealTimeScheduler callbacks.