envir: Removed isModule() guards from recording a component method call begin.
I think this is a leftover from ancient times when we had to have a module to get its id.
The old code didn't work for channel method calls, because only the component method call end
was recorded. The resulting eventlog file could not be read by the sequence chart tool.

Also: Avoid creating imbalanced begin-end entries for from==nullptr.