qtenv/loginspector: Fix a crash while setting/clearing reference time
This method is used, among others, to find where the caret should be drawn.

Before the "reference time" feature was added, it was basically impossible
(or at least _very unlikely_) for the caret to be inside an ANSI escape
sequence. Even though it's stored as an index into the raw line content
(incl. escape sequences), it was always always placed such that the index
was computed back from an index in the stripped version of the line -
which never put it into an escape sequence.

Now, it's possible for the "effective content" of a line to change underneath
the caret (setting/clearing the reference time does that), and it's position
is not adjusted when this happens. Therefore, this code can now find itself
having to deal with a situation that it wasn't prepared for. Until now.

The crash itself was the direct result of passing in a negative string length
to the QString factory method before measuring its displayed width.