Revert "sim: trim unused result filters / recorders"
This reverts commit 47eab84bb3290f8e976d1f631c9fa402cf873eba.

Reason: The patch made libsim depend on libenvir, which caused linker
errors on Windows and macOS. One possible solution would have been
to move the guts of isRecorderEnabled() into libenvir by adding
isVectorEnabled() / isScalarEnabled() methods to cEnvir. However, changing
the cEnvir interface does not seem warranted, given that the expected
speedup of simulations from the patch is is not that significant.