samples: fix osg-indoor error: Removed m unit from playground parameters
Using $variable substitution in display string accesses the parameter
without specifying a unit. This is now throwing an error
(accessing parameter defined with a unit, but without specifying the assumed unit
when retrieved).

This is unrelated to the actual sample (OSG), so we just remove the unit and
assume that everything is in meter.