ide: fix debugger not launching on Windows/msys
On windows, eclipse cannot launch the opp_dbgmi script directly as it is
not aware how to deal with shebangs (as eclipse is using standard windows
CreateProcess calls).

Calling as 'bash opp_dbgmi' also leads to an error because Windows has a
bash.exe in windows\system32 which starts the default WSL instance
instead of the msys bash executable.

As a workaround we use `sh ${opp_root}/bin/opp_dbgmi` which can work also
on Windows and it does not interfere with the Windows bash executable.