ide: add quotes around command line options if a quoting is required. (like it contains space or semicolon)
The command line was printed to the console before the simulation started,
but copying it verbatim may end in an invalid command if an argument contains a semicolon (like ned exclusion lists)
as the shell would split the command-line along the ; chars.
Quoting properly these arguments prevents this incorrect behavior.