prod: fix #1209: install.sh does not install dependencies on Fedora/openSUSE
package managers usually report the existence of available updates by returning
a specific non-0 return code. The install script stops on the first
error so it stopped the execution as soon as the package manager reported
that there are updated packages available.

Remove the explicit package manager updates as it does not belong here anyway.
The user must explicitly update the package manager's package database
if they wish to start with a fresh database. Also updating the database in
the script would prevent a docker image to be repeatable if the install.sh
script is used inside a docker recipe.