ide: use an older version of GEF because of an issue on Linux/Waylad: REVERT this patch after 2024-12
because of the issue at: https://github.com/eclipse/gef-classic/issues/568
we have to use a nightly version of GEF classic. The issue is already
fixed in the head of the gef-classic repo, but the eclipse 24-09 release
already contains the version with the issue, so until we don't upgrade
to eclipse 24-12 we should keep this patch.

After that date PLEASE REVERT this and remove the nigly gef repo URL from ui/pom.xml