ide/scave: Make the "Zoom to Fit Data" action on native line plots accurate
This is a bit of a hack, to help setZoomedArea() be a bit more accurate.

When the "Zoom to Fit Data" button was pressed on a maximally zoomed out
native line plot, the shown area was a bit smaller than what was intended.

The zoom level was calculated for the viewport size with the scrollbars
not being visible - but at the end, they were set to be visible,
shrinking the final shown area a bit.

We make an educated guess about the final visibility state of the
scrollbars, and set them so early, to start the layouting process
with the final viewport size.

In addition, the target view area is stored throughout the layouting
process, to make sure that the changing Y axis tick label sizes don't
shrink the visible X range.