qtenv: Fix animation of broadcast sends with broadcast animation turned off
Previously, only the first send was properly animated, because all of the
holds started at the same time, therefore the rest of them already had
"expired" right when the second one was about to start. Decoupling
sequence init from begin, and not calling init() on the send animation
right away fixes this. The lines are immediately shown, but the holds
for the message animations are not all started at the same time.
See test/anim/broadcast and test/anim/animation2.